+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 ### Fixes
 - Overriding an animation with same name linked to a subscene being removed from it
 - Hurt bobbing when changing camera angle (I couldn't this bug, but I assume I fix it I hope)
+- Cutscene editor playhead staying still when pressing play